toolkit/components/captchadetection/tests/mochitest/test_hcaptcha_pc.html

<!doctype html>
<html>
  <head>
    <meta charset="utf-8" />
    <title>hCaptcha Manual/Puzzle Completion Detection Test</title>
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" href="/tests/SimpleTest/test.css" />
    <script src="GleanTest.js"></script>
    <script src="CaptchaTestingUtils.js"></script>
  </head>
  <body>
    <p id="display"></p>
    <div id="content" style="display: none"></div>
    <pre id="test"></pre>

    <script>
      add_task(async function () {
        await GleanTest.testResetFOG();

        const imagesFrame =
          await CaptchaTestingUtils.createIframeAndWaitForMessage(
            "hcaptcha_puzzle.html#frame=challenge",
            document.body,
            "ready"
          );

        imagesFrame.contentWindow.postMessage("show-puzzle", "*");
        await CaptchaTestingUtils.waitForMessage("showed-puzzle");

        const checkboxFrame =
          await CaptchaTestingUtils.createIframeAndWaitForMessage(
            "hcaptcha_checkbox.html#frame=checkbox",
            document.body,
            "ready"
          );

        const metricSet = CaptchaTestingUtils.waitForMetricSet();
        checkboxFrame.contentWindow.postMessage("display-checkmark", "*");
        await metricSet;

        const puzzleCompletedCount =
          await GleanTest.captchaDetection.hcaptchaPc.testGetValue();
        is(puzzleCompletedCount, 1, "The puzzle completion count should be 1");

        await CaptchaTestingUtils.clearPrefs();
      });
    </script>
  </body>
</html>
